Perform version conflict resolution when selecting
This change consolidates all of the logic to choose a particular module
version for the current set of selectors, when a new selector is discovered.
Previously, version-range merging was done eagerly, but actual conflict
resolution was deferred until later in graph resolution. This disconnect
led to a number of selector-ordering issues and made it difficult to
consider all known version constraints when making the choice.

For now, module version selection is still done in a pairwise fashion:
only the current selection and the newly discovered candidate are
considered. But this new algorithm will make it possible to consider
all possible candidates when resolving, allowing better choices to be
made in the presence of `reject` constraints.