Construct reject failure reason when validating graph
Instead of building the failure reason during conflict resolution
when we find a rejected component, we now flag the component as rejected
but only build the message later when actually validating the graph.

This required a change to the message builder, so that it constructs
a failure message based on the edges and selectors in the graph, rather
than attempting to build this from the components provided to conflict
resolution.