Change the `TaskContainer.createLater()` method to return a `Provider<T>` that references the task. The task is created and configured when this provider is queried.
Change the Java base plugin to use this to wire in the `processResources` task for a source set into various places, rather than looking up the task by name immediately after creating it. The task is still queried eagerly during configuration, and so created, because the things that receive the task currently are unable to receive a provider.