Change the Java base plugin so that it does not attach the `classes` task for each source set as the "build" task for the associated binary.
This is a breaking change, but this breaking change is on the software model view of the Java binaries and so is unlikely to have any impact. This breaking change can be fixed with some more effort to avoid realizing the classes task in the software model rules, if we need to.