Avoid accessing `ResolutionStrategy` for non-root components
To determine if a configuration should have dependency locking enabled,
we inspect the ResolutionStrategy for that configuration. However, we
were needlessly inspecting this for configurations in referenced
projects, leading to added contention.

With this change, we only inspect the resolution strategy for
configurations when building the component metadata for the root
component, not for any referenced project components.