Change `DomainObjectCollection` implementations so that actions added using the eager `all()` and `withType()` methods receive elements as soon as the element is defined, regardless of whether added using a provider or directly, for backwards compatibility.
This change means that the `compileJava` task is created eagerly during configuration due to the presence of these kinds of actions used in the Java and Java base plugins.

Add a method to `DomainObjectCollection` to allow elements to be configured lazily, as required. This method isn't used by the Java plugins yet.