Consider all reject constraints when resolving dynamic versions
When resolving a component selector to a component identifier, we now construct
a unified 'allRejects' constraint that allows the correct resolution of
dynamic versions. If all versions matching a dynamic version are rejected,
then the highest rejected version is chosen. Checking for rejected versions
in the resolved graph is done in a subsequent step, after conflict resolution.

When the set of reject constraints changes there may be a need to re-resolve
a particular component selector. Efforts are made to reduce these situations
by comparing the previously resolved version to the updated set of constraints.