Fix flakiness with MSBuild and plain console
The plain console prints the build failure on stderr.  MSBuild
runs gradle in a subprocess, but then joins stdout and stderr
from that process onto stdout.  It seems that it does some
buffering and sometimes stderr gets jumbled up with stdout in the
output.  This can cause the output scraping to fail.

This changes `MSBuildExecutor` so that output from the
gradle command is redirected to separate stdout and stderr files
and then we parse the gradle build output for verification from
there.