Increase directory cache cleanup frequency from 7 days to 24 hours
With this change it now makes sense to set the cleanup timeout to small numbers like a few days. Previously we'd only check for outdated entries in local directory caches once a week, which meant leaving unused files in place for much longer than the user would have expected.

This change currently affects the cleanup of the build cache and the cache used to store source dependencies.