Make plugins.withId more efficient in the common case
Use the cached plugin descriptor if the plugin was
loaded by the classloader of the plugin registry that
it was added to. This is the case 90% of the time.
The only common case where this is not true is when
a plugin is loaded by an init script or script plugin.