Change exception handling when phased action fails
Exceptions in BuildActions of a PhasedAction are now immediately thrown
and the build imediatelly fails. This makes sure that if an action fails
the remaining steps of the build will not uselessly be executed.

Exceptions are unwrapped in ProviderConnection so the correct
information is sent back to the TAPI client.

This commit also addresses other review comments in the PR:
- Removes unnecessary @since annotations in methods
- Adds a test (ignored for now) making sure that default tasks are not
run when no tasks are specified by the user

Signed-off-by: Lucas Smaira <lsmaira@google.com>