Add 'rejected' status to `ComponentIdResolveResult`
We now consider all reject constraints when resolving version constraints to a
component identifier. Previously, this knowledge was discarded and we then
re-evaluated the reject constraints to see if the resulting component was indeed
rejected.

With this change, the `ComponentIdResolveResult` retains the knowledge that a
component that was 'selected' for a version constraint is in fact 'rejected'.
This is done for both dynamic and static version constraints, removing the need
to re-check the reject constraint after selection.