Read dependency attributes in Gradle module metadata
This commit adds support for parsing dependency attributes in Gradle module metadata.
It is not yet making use of that information, it's only making sure that the parser
is able to parse and propagate it. It doesn't touch the serializer either, which is
going to be done in a separate commit.

Module metadata file version has been bumped to 0.4 as part of this work.