Add test fixtures for publishing dependency attributes
This commit introduces a way to declare attributes on dependencies in published Gradle
metadata, using test fixtures. This makes it possible to consume dependencies on attributes
and make sure that they are honored the same way as if they are defined using component
metadata rules (a test cases has been added for that purpose).

Publication of attributes on dependencies is not happening yet.