Make plugin lookups cheaper
Use the classloader scope of the registry as often as possible
instead of using a potentially non-caching classloader.

Move classloader caching to the root classloader scope.
This is more consistent with how the child scopes work.
Also the caching is not needed for the other clients of
the classloader registry.