Add intro and example scenarios to Java testing chapter
I now cover some for the basics of testing so that readers have a foundation
before going into the more specific topics. I have also added some common
scenarios that users encounter, with examples of how to tackle them.

Finally, I've added an integration test for all the samples in this chapter.