Implement matching of component attributes during dynamic version selection
This commit adds support for matching the component attributes during the selection
of a version, when the selector is dynamic. Before, component-level attributes were
ignored, meaning we could select a version which wouldn't have any variant matching
the consumer attributes. With those changes, it is possible, for example, to have
a version, 1.3, which doesn't match the consumer attributes, and an earlier version,
say 1.2, which does. In that case, 1.2 would be selected instead of 1.3.

The consequence of this change is that we now have to fetch metadata for components
that are selected during dynamic version selection, in order to check if they match
the component attributes. This used to be the case for status selectors (`latest.release`)
but not for version ranges (`[1.0, )`).

It's also worth noting that if the selector is dynamic, we will perform attribute
matching at least twice (possibly with different attribute sets):

- once during version selection (for each candidate, check if its attribute match)
- then once during _variant_ selection (in which case each variant may have additional
attributes)