Inline `isTransitive`
The dependency resolution engine heavily relies on calls to `edge.isTransitive`, which in turn
delegates to `metadata.isTransitive()`. Because we have more than 2 implementations, the JIT
cannot inline the calls to `isTransitive`, therefore we execute virtual method calls which are
clearly visible in hotspots in profiling data.

This commit inlines all the `isTransitive` fields, which will add a byte to several objects,
but will in turn make dependency resolution faster.