Update lists of tasks in Java Plugin chapter
I've made the styling of the definition lists consistent with other chapters
and updated some of the content that was looking a little long in the tooth.
I also removed italicisation of the first sentences of definition list
paragraphs because it was overkill and a little hard on the eyes. There was no
real need to emphasise those pieces of text and long sentences of italics can
grate.