Remove model rules from java-base plugin
The java-base plugin has a bridging layer that
makes SourceSets available as binaries in the
software model. Using the software model for
JVM projects was never fully implemented and is
now discouraged.

The plugin now only creates those bridged binaries
if another plugin or build script uses the software
model.