Cache version selectors parsed from version scheme
During dependency resolution, we need to create a lot of version selectors. In practice,
we often create the same version selectors again and again for the same version strings.
This commit introduces a caching version selector scheme, at build scope, that allows us
to share instances of version selectors, and therefore reduce the time spent in `getSelector`
as well as reducing garbage.