Make pending detection more accurate (#5105)
This is an attempt to fix https://github.com/gradle/gradle-private/issues/1225

Previously, the failed test tried to wait for the changes to be picked up in continuous by waiting for the keyword "Pending changes detected", which is not accurate enough. We should wait for "Pending changes detected" and the following build completes.