Add support for write-only properties to DSL Reference generator
Prior to this commit, metadata for properties that only have setters but no getters was collected. However, when generating the DSL docs, a NPE was thrown when no getter was present. Now, write-only properties are fully supported, their Javadoc comment is adjusted ("Sets the thing" -> "The thing") and they are rendered using a "(write-only)" suffix.