Improve error reporting in case no matching dynamic version is found
This commit improves rendering of errors in case resolution fails because
all versions in a dynamic range are evicted, and that at least one of the
evicted versions was evicted because of attribute matching. The error will
now report the attributes on each tested version, as well as the requested
attributes.

For this, the module not found exception has been updated to carry more
context, and now makes use of the tree formatter for consistency with other
exceptions in the codebase.