Fix usages of spec.sourceFiles.asFileTree
Now that we use an immutable file collection for `sourceFiles`, calling
`sourceFiles.asFileTree` has an actual effect. Before using immutable
file collections it was expected that each entry in `sourceFiles` has
the right corresponding relative path.

This only has effect on incremental compilation for now. The solution
is to match on class name manually instead of using "magic" include
specs. This avoids the creation of a new file tree altogether.