Fix typed attributes on dependencies triggering an assertion error
The attributes that we write in the serialized graph must be desugared, similar as if
they were read from an external module descriptor (Gradle metadata). This was done for
variant attributes, but now that we can set attributes on dependencies too, they must
be desugared too. Without this fix, attempting to use a dependency attribute with a
strongly typed attribute would result in an assertion error in the serializer.