Also desugar component selector when serializing errors
The component selector also needs to be desugared when we get the requested
selector from the edge state, in case of a failure to resolve. If we don't
do this, the internal map of errors we get contains a map from typed attributes
to errors, then when we read back, we get an untyped selector, triggering
a null pointer exception.