Add support for version lister on Ivy repositories
This commit introduces the concept of custom version lister. Whenever
a dynamic version of a module is requested, Gradle has to perform a
lookup on a remote repository to list the available versions of this
component. With this change, it is now possible for a build author to
provide a custom version lister. This can be used to solve different
problems:

- it becomes possible to have a single remote resources that lists all
versions of all known module, for example
- avoid hitting a remote repository if we know a module will never be
found there

Note that for the latter, if the version is _fixed_, the version lister
will never be called so we will reach the remote repository in any case.

For the purpose, an interface, `MetadataSupplierAware` has been extracted
and is for now only used by Ivy repositories, because the latter was the
only one providing component metadata suppliers.

With this change it will become possible for Maven repositories to
support that too.

It's worth noting that the goal of this separation of concerns is to make
it possible for a user to either supply a version lister, a component
metadata supplier, or both. And if using both, a shared cache can be used
to avoid reaching a repository multiple times. This means that in theory
it should be possible to perform a single network request to both list
the versions of a component **and** get the component metadata necessary
to perform the selection of the component.

This commit does _not_ declaring the attributes of a component yet.