Ignore unused non-incremental processors
Compilation can be incremental even if there are
non-incremental processors on the processor path
as long as those processors are never used. This
is quite a common scenario since incremental and
non-incremental processors may be packaged up
together or a non-incremental processor may be
accidentally leaked as part of a dependency.