Allow dynamic incremental processor registration
Some annotation processor may not know whether they
can be incremental until runtime. The Google AutoValue
processor for instance has an extension mechanism and
needs to check whether all extensions are incremental
first.

This supersedes the static discovery (based on META-INF)
with a dynamic processor-option based approach. This has
the added benefit that we can easily add more options in
the future.