Make incremental annotation processing lenient
Instead of failing when the incremental contract is violated,
Gradle will now do a full recompilation instead, warning the
user about the violation that caused this.

This should make it much easier and safer to mark certain
processors as incremental. For instance, the MapStruct
processor has a rarely-used feature that generates a resource.
Previously it couldn't have been marked as incremental because
resource generation would have been an error. Now it can be
marked incremental and if a user uses the resource-generating
feature, it'll simply result in full recompilation.