Fix component metadata rules not applied on user provided metadata
This commit fixes the fact that component metadata rules were not applied
on "user provided" metadata. User provided metadata comes from a component
metadata supplier, and is only used during selection. If, for whatever
reason, metadata rules needed to change the "shallow" metadata at this
point, then we would fail resolution because those rules were not applied.

This is done by creating an adapter which is just going to apply the rules
that make sense on shallow metadata, which is just for now the status and
the status scheme, but is going to include more in the future (in particular
the component attributes).