Add support for mutating component attributes in a component metadata supplier
Attributes can be used during selection of a dynamic version, however there wasn't any
way for a component metadata supplier rule to provide the attributes of a component.
This commit fixes this by exposing `getAttributes` and `attributes` methods in the
component metadata builder, so that the supplier can feed those before selection
happens.