Make sure component metadata supplier is used whenever custom attributes are requested
This commit makes sure that the supplier is going to be called whenever the consumer
wants custom attributes (that is to say not only the "status"). This shows that we
are also going to call the supplier for version selectors which normally don't require
metadata (sub-version selectors), but are now insufficient because the attributes are
part of the component metadata.