Better cycle breaking algorithm
The new algorithm first adds non-breakable edges one-by-one. If we find an edge that creates a cycle, we fail the build.

Then we continue with breakable edges, and if we find an edge that would create a cycle, we simply skip it.

For each edge we need to check the parents of its source node to see if it would would introduce a cycle. Given a well-balanced build the algorithm should take O(e * log e). In the worst case of a linear build it would take e * e / 2.