Revert changes to SourceToNameConverter
The performance improvements made this class
inconsistent with other classes doing similar
jobs like OutputToNameConverter.

Since incremental builds rarely change a significant
amount of the code base, it is unlikely that this will
actually become a bottleneck. The reason for the perf
improvements was a different usage of this class,
which we have removed in the meantime.