Remove download message in play tests to avoid flakiness (#5267)
This is an attempt to fix https://github.com/gradle/gradle-private/issues/1238.

This flaky test can be reproduced by removing `<projectDir>/intTestHomeDir/worker-1/caches`.

If caches is empty when executing this test, extra `Downloading http://xxx.jar` will appear in test log, failing the test with unmatched log failure.

This PR removes the downloading message from test log.

In addition, this PR fixes an issue in `GroupedOutputFixture.java`: previously, if the test output is `\n\t\tOutput\n`,  it will be trimmed to `Output` - this is not right. Only empty lines should be trimmed, so the correct trimmed output should be `\t\tOutput`.