Fix flaky test stopScheduledExecutorThrowsExceptionOnTimeout (#5268)
This is an attempt to fix https://github.com/gradle/gradle-private/issues/750

This flaky test's intention is:

- Submit a thread which blocks 750ms to the executor
- Close the executor with 200ms timeout
- Assert an exception to be thrown

The problem is, when the executor start closing, the block thread might not start running. Therefore, the executor can close without any exceptions.

This PR adds a `CountdownLatch` to coordinate these actions. The executor will not be closed until block thread starts running.