[WIP] Use legacy reporting name when no ClassSource is available
This commit removes the requirement for JUnit Platform Test Engines to
return a `ClassSource`. Instead, the test tree node's legacy reporting
name is used as a fallback.

In addition, the TestExecutionListener implementation now calls the
TestResultProcessor infrastructure directly which immensely simplifies
its implementation and allows for reporting events for all nodes in the
test tree which in turn enables better reporting of failures of
composite nodes, e.g. when an engine has a configuration issue.

Resolves #4912.