Run init scripts from the client distribution
The recent change to start parameter serialization
had omitted the Gradle home directory, leading to
Gradle chosing the distribution that the daemon was
started with over the distribution that the client
wanted. This is now fixed.

This also fixes another issue where the Gradle home
was not passed on to nested builds like buildSrc,
which lead to the same problem for them.

Gradle#getGradleHomeDir will now correctly return
the home dir that the client selected, not the home
dir that the daemon was originally started with.

The test veryfing all this has been fixed so that
it asserts that the daemon was reused. Previously
it was just creating new daemons and thus passing.