Make UriNotationConverter faster on Windows
Windows paths can look very much like a URI,
which is why we need to disambiguation between them
by looking at the file system.

However, this is only necessary if the notation starts
with what looks like a URI scheme. In the vast majority
of cases, the notation converter is called with relative
paths that don't have this problem. This change makes
that common case much faster.