Hoist special handling of PCH files into IncrementalNativeCompiler
Instead of passing around a map of source file to list of literal includes,
we pass a set of source files known to be PCH safe.

When no PCH is used, we don't populate the list.
When a PCH is used, we populate the list of files with any source file
that includes the PCH as its first include.