Add a overview chapter for publishing
This commit extracts the common parts of publishing from the Maven Publish and
Ivy Publish Plugin chapters and restructures them into an overview that should
be easier for newcomers to understand.

One result is slimmer plugin reference chapters. I have also added publishing
sections to the Java and War Plugin chapters detailing the components that they
define.

Finally, I have updated the formatting in the War Plugin chapter so that it's
consistent with other plugin reference chapters.