Fix "forced" flag being sometimes ignored
If the first direct dependency we see for a specific module version (g,a,v) has forced=false, then
if another first level dependency (g,a,v) with forced=true is added later, we ignore the fact that
the second one has forced=true. This case was lost during refactorings, due to insufficient test
coverage.

Fixes #5364