Fix incorrect conflict resolution when a cycle exists with root component
Whenever the root component was added back through a dependency cycle, we
introduced a conflict, which was solved by the latest module conflict
resolver _unless_ `failOnVersionConflict()` was used. Instead, we shouldn't
even try to resolve a conflict because there's none. This commit fixes the
problem by checking if the root component is already added.

Fixes gradle/gradle-private#1268