Allow at most 3 parallel performance coordinators
While having the performance agents busy is important, it seems to me
that the average time to complete each coordinator went up due to all
performance agents occupied.

Looking at the Teamcity statistics the average running time of the
coordinator build went from 80 minutes to 155 minutes. I think that
justifies re-instantiating the limit.

Especially when the pipeline is busy, we should limit the number of
parallel performance test coordinators running.