Use cross-build cache for component metadata supplier caching
This commit introduces a cross-build caching mechanism for component
metadata supplier rules, making them effectively reusable accross
builds, and has an in-memory facade, to make it more efficient for
multi-project builds (the condition being that the same rule class
is used by all projects).

Cache invalidation is handled by the `CachePolicy`, meaning that we
honor the `mustRefreshModule(...)` calls, making it in particular
sensitive to `--refresh-dependencies`.