Clean up least recently used entries in artifact transforms cache
The files in the transforms cache are now automatically cleaned up based
on a least recently used strategy. The structure in the cache is as
follows:

 - files-1.1
   - artifact-a
     - hash 1 (*)
       - hash 1.1
       - ...
     - hash 2 (*)
     - ...
   - ...
 - metadata-1.1

The cache now keeps track of files accessed inside the first hash (*)
level by touching the corresponding directory. The cleanup action then
deletes all such directories that have not been accessed in the last 7
days.

Issue: #1085