Fix concat for attributes with same name, different type
When 2 containers contain attributes with the same name and different
types, the `concat` and `safeConcat` methods of `ImmutableAttributesFactory`
now treat these attributes as equal for the purposes of merging.

Previously, these methods would produce an `AttributeContainer` 
containing both attributes. 
This was a violation of the `AttributeContainer` contract which states:
"It is not allowed to have two attributes with the same name but 
different types in the container"