Allow a build operation notification listeners during continuous build
This accidentally worked < 4.7. In 4.7 the notification bridge became cross session scoped in order to support bridging the logging output to progress notifications. This broke what was accidentally working, by retaining the notification listener across builds of a continuous build.

This change restores the accidental support by explicitly clearing the bridge state after each build of a continuous build, allowing a new listener on the subsequent build.