Implement discovery of rule inputs for injected services
This commit replaces the not-so-nice service "snapshotting" with a proper
implementation of input discovery for services. The idea is that rules can
have injected services. Those services are usually blackboxes, but they
may actually do work which is "expirable". Typically, a rule which fetches
an external resource will need to re-execute once the resource is outdated.

The problem is that the service is called _at execution time_, so we only
know about the actual rule inputs once it has effectively been executed.
To implement this capture, we need services to declare that if called,
they will participate in some inputs. In that case, we record a "query"
to the service, as well as the fingerprint of the answer to the query. Then
the result of the rule is cached. Once we fetch the result from the cache,
we ask the service if the result for the cached query is up-to-date. If not,
we discard the cache entry, otherwise, we consider it's up-to-date.

This removes the need for "snapshotting services" as we consider that a
service which doesn't declare itself as contributing inputs has no side
effect (it will always return the same value, independently of time).