Fix registration of module registry
The service registry method needs to return the concrete type
of a service in order to discover all of its provided interfaces.
In this case the module registry was not being picked up as an
immutable file location, because that interface could not be
discovered from the method return type.