Use daemon by default when forking from embedded executor
The Play! integration tests always fork a new process
so that they can cancel it. The embedded executor was
using no-daemon mode when asked for a forking handle.
This is now changed to use a daemon instead, greatly
speeding up the Play! integreation tests in embedded
mode.

I stumbled over a bug in our module registry that made
it impossible to run anything in IntelliJ in embedded
mode that used a type from the "plugins" project.
I suspect that some "requireDistribution" usages might
be just because of this bug. We should definitely check that.

There was also some inconsistency in how the embedded executor
handled jvm args and environment variables when it forked a
process. This went unnoticed because it wasn't using the daemon
and the arguments it was missing were to work around issues in
daemon mode. This is now fixed by extending the daemon executor.