gradlew: use `readlink` to read a link instead of attempting to parse `ls`
Parsing `ls` is notoriously fragile and should be avoided when possible.
Let's use `readlink` (part of coreutils, just like `printf` which is
already used in this file) to read links instead of a fragile parsing of
`ls`.

The current code breaks if any part of the link or its derefs contains
`-> ` (uncommon, granted). You can test that breakage with:
$ ln -s 'bar -> baz' foo

Signed-off-by: Eric Engestrom <eric.engestrom@intel.com>