Make dynamic annotation processors explicit
Calling Processor#getSupportedOptions() is not allowed before
calling Processor#init(). However, we cannot decorate the environment
if the processor is non-incremental, because some processors
expect to be able to cast the environment to its implementation
type.

To fix this, we now require processors to explicitly declare
in their META-INF directory that they may be incremental, but
need to decide at runtime whether they really are. These "dynamic"
processors need to be fine with Gradle wrapping the processing
environment.