Add dedicated DSL to customize Ivy descriptor to ivy-publish plugin
This commit adds a type safe DSL for customizing the generated Ivy
module descriptor of an IvyPublication to the ivy-publish plugin:

  descriptor {
    license {
      name = 'The Apache License, Version 2.0'
      url = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
    }
    author {
      name = 'Jane Doe'
      url = 'http://example.com/users/jane'
    }
    description {
      text = 'A concise description of my library'
      homepage = 'http://www.example.com/library'
    }
  }

Only interfaces are exposed as part of the public API, all of them are
prefixed with `IvyModuleDescriptor`. The exposed properties make use of
the Provider API.

In addition, the new DSL is documented in the User Guide, DSL Reference
and Release Notes.

Resolves #5193.