Make `failOnVersionConflict` non-fatal on demand
Before this commit, `failOnVersionConflict()` was triggering a failure
during dependency resolution as soon as a conflict was discovered. This
commit changes this so that conflicts are resolved as usual, but a
validation step triggers an error after resolution if a conflict is found.

The `ResolutionResult` API has been updated so that a consumer can
set an alternate error handler, in which case whenever the resolution is
triggered, we can intercept the error and do something. In case of the
dependency insight report, this now allows not failing the build if
there's a resolution error due to the fail on conflict flag, but instead
show a resolved graph with a warning that a conflict happened.