Support dependency insight report when lock file is out-of-date
Whenever a lock file is out of date, we need to make sure that we can still execute
the dependency insight report. Without this change, any attempt to execute the
report task when the lock is out of date is going to fail. This makes the behavior
of the report consistent with the fail on version conflict flag.