Collect all rejections during dependency resolution
This commit collects all rejections during dependency resolution,
so that we can properly show them in the dependency insight report.
This highlights some consistency/readability issues, like when a
rejection originates from a constraint. For this reason this commit
does **not** pass tests, and is just the first step towards better
rendering.