Collect all dependency reasons
This code fixes the problem that we only keep the first dependency reason
that we find when resolving. If it was a dependency without reason, then
even if a transitive dependency had a custom reason, we wouldn't show it.

Now, we properly collect all reasons during resolution, then show them in
the dependency insight report. It's worth noting that the output of the
report has been changed to always include a header, so that we have the
details for all dependencies which match the predicate.