Add rendering of failures to dependency insight report
This commit adds support for rendering failures in the dependency insight
report. Before this, the only message you would get was `FAILED`, which
forces to implement a task to show the actual error, which can be a bit
painful sometimes. Instead, errors are now reported as extra details in
the header.