Allow dependencies on a project with auxiliary publications
Users commonly publish their test fixtures alongside the main
component. Until now, depending on such a project would result
in an error during publishing, saying the Gradle can't decide
which of the publication coordinates to depend on.

Ideally we would model test fixtures as a first-class component
or provide an extensible API for users to define their own components
and their relationship. But to unblock this very common use case
quickly we decided to use a simple heuristic for now:

If a project only has one publication with a component, that publication
is the one we use when generating dependency declarations. The other
non-component publications are assumed to be auxiliary.