Remove flakey and previously not running tests
This test approach does not work. This test has been around for a while, but was missing `assert`s so wasn't actually testing anything. Turns out it is incredibly flakey due to the inherent non determinism from the console renderer debouncing updates (i.e. the sequence of writes to the console is non deterministic).