Rebaseline performance tests
To account for a regression across most tests,
which was caused by a bugfix for lazy task creation.
Lazy task creation was only that fast because of
that bug, so this new performance level is the
expected one. We can definitely improve it further,
see https://github.com/gradle/gradle-native/issues/678