Rebaseline Android performance test
This got a bit slower now that we clean javac's zip file cache.
This fixed a bug where we would keep JARs open for the lifetime
of the daemon, preventing them from being deleted on Windows.

We could make this cheaper by only clearing the cache once,
after the build is finished.