Lazily create component selection reason descriptors
With the gathering of all component selection reasons, we introduced a performance regression
related to the creation of the descriptors. Most of them are due to strings being eagerly
constructed. However, not all of them are necessary: only the ones for the components which
will actually be selected in the graph. For this reason, this commit changes the implementation
of the component selection reason to use `Describable` instead of `String` for descriptions.