Continue executing tests if irreplaceable security manager was installed
Previously, TestWorker failed to handle a SecurityException thrown when
it called System.setSecurityManager(null). This commit makes TestWorker
properly handle any SecurityException that is thrown, allowing tests
that install an irreplaceable security manager to exit cleanly.

In addition, the TestWorker now only attempts to reset the
SecurityManager to the previous value if it was changed.