Ignore --foreground tests on IBM JDK
The foreground daemon uses whatever environment
variables were set when it was started, as it has
no "client context". On IBM JDK there are a bunch
of extra environment variables by default, which
interfere with subsequent matching of this daemon.

The whole --foreground feature seems questionable,
but that's a decision for another day.