Inject Gradle version into installation chapter
I have added a `gradleVersion` Asciidoctor attribute to the docs build. The
installation chapter now uses that to display the appropriate Gradle version
in various locations.

In order to get this working for the code blocks, I had to add

    [subs="attributes"]

before each of the code blocks that referenced the `gradleVersion` attribute.