Use separate cache for file access time journal
Storing a file access time journal inside each cache causes a lot of
contention in particular with caches that require a lot of exclusive
access like the artifact cache.

Instead, a new `journal-1` cache that is managed by a user-home-scoped
service is now used to keep track of file access times for all caches
that want to use it.