Normalize `ModuleIdentifier`
This commit reworks the `ComponentModuleIdentifier`/`ComponentModuleSelector`/`ModuleVersionSelector`
classes to use `ModuleIdentifier` under the hood, instead of storing denormalized strings. This has
the advantage that we can reduce the use of the module identifier factory, which is called very
often during dependency resolution. Sharing instances reduces the need for conversions, and makes
comparisons faster.