Stop publish plugins from overriding values set by the user
If stable publishing is not active and the user uses
`tasks.withType(GenerateMavenPom)`, we would overwrite
the destination dir value set by the withType action.
This is because creation closures in DomainObjectContainer
are run after previously added actions, while ModelMap
would run the creation closure first.

We can remove this special casing in 5.0. Generally it would
be good to have a default value concept on Properties.