Fix edges being incorrectly reintegrated into result
It was possible for an edge to be reintegrated in the graph even if its
source node wasn't selected. This would result in an inconsistency when
writing, then reading the generated serialized graph. In particular,
when a node had a replacement, an that another edge in the graph caused
an upgrade of the source dependency, the older version of the source
node could have been seen as selected when it wasn't.

Fixes #4785