Treat `latest.xxx` differently when dependency locking is in action
This commit adds special handling for the "latest" selector in case
dependency locking is used. This is technically a workaround so that
the expectation from a user point of view matches the principle of
least surprise. In particular, before this commit, if the --update-locks
flag was used on module A, and that module B had "latest.xxx" as
requested version, then we would also upgrade the version of module B,
because "latest.xxx" means "the latest xxx", not "the latest xxx matching
other requirements" (where other requirements = locked version).

Fixes #5531