Use a daemon thread for log message dispatch, to avoid leaving non-daemon threads running in a tooling API client when the client has finished with the tapi.
Currently, there is no method that the client uses to signal that it has finished with the tapi, so we have to assume that the client will exit at any time. It would be better to make this an explicit call so that these threads and other resources can be released when the client is done with them.