Canonicalize relative wrapper distribution URIs
Prior to this commit relative file URIs used in wrapper.properties
were not canonicalized which led in some cases to lots of duplicate
distributions in wrapper/dist because each URI had a different hash.
Now we attempt to canonicalize the file path before converting it into
and URI. In case that fails, we fall back to the old behavior.

Fixes #2844.