Use inception time of journal as fallback
Prior to this commit the file modification timestamp was used as
fallback when the last access time was unknown. That lead to a lot of
files being deleted when first using the new Gradle distribution that
includes cleaning up of shared caches. In particular for dependencies
that have not been updated in the last 30 days and were not used in the
first build, lots of files would have to be redownloaded

This commit persists the inception time of the file access time journal
and uses it as the fallback for unknown files. This way, less files will
be cleaned up initially.