Set resolution result even in case of failure
Whenever resolution validation fails (in which case the `failed` method is
called), we still want to be able to show the dependency resolution result,
which is a valid graph (in particular in build scans). This commit fixes
this by calling both `failed` (to indicate that the result of resolution is
a failure) and `setResult` (to indicate that despite the failure, there is
a valid, safe, dependency graph to show).