Decouple Play and Twirl (#2062)
This change fixes support for Play 2.6 which is currently using the Play 2.3 import list.

This also theoretically makes it so that you can use varying versions of Twirl with a given version of Play, which is something that makes upgrades between Play versions easier, allows for pulling in bug fixes for Twirl separately, and is something supported by the SBT plugin.

This changes the imports to a List from which the concatenated string version is then constructed. This change is necessary because Twirl 1.3 expects a List. (Well maybe not strictly necessary as I could explode the concatenated string version into a List, but that wouldn't be as clean)