Do not generate multiple progress start and complete events for each task as it executes.
Set the relevant fields on the start event associated with the task execution build operation instead of generating a second progress event to carry this information. Also change the console grouping mechanism to use the event _description_ instead of the event _logging header_ to render the group header. This will allow the logging header concept to be removed in a subsequent change.