Fix build failure
Changing the DSL version broke the test build:
https://builds.gradle.org/viewLog.html?buildId=14019951

Turns out that the existing source configuration specified a wrong
source root which lead to a weird compilation error. The fix was
to use `.` as a source root.

Also, I copied over the Maven configuration from a fresh TeamCity
DSL project to avoid further problems.