Fix potential NPE in performance report
Since our query for test executions and measurements are
based on their start time, it can happen that a test execution
is skipped because it was too old, but one of its measurements
was started later and is included in the result. When this happens
we cannot fine the corresponding execution, resulting in an NPE.

We now ignore measurements for executions that were skipped.