Make tests robust against additional build operations
Prior to this commit, a lot of tests asserted that there's a single root
build operation. This assertion was too strict and did not account
for potential additional root build operations such as the recently
introduced cache cleanup for the artifact and artifact transform caches.
Now, `BuildOperationsFixture` provides a new `root()` method to get the
root for the supplied `BuildOperationType`.