Simplify HttpClientConfigurer
Prior to this commit, we formatted a string from the header name and
value, parsed it back into name and value in `HttpClientConfigurer` to
put it into a `BasicHeader` in `HttpClientHttpHeaderCredentials`. Now,
we just pass through header name and value.

In addition, `getCredentials()` in `HttpClientConfigurer` has been
inlined to move the check for which implementation class is being used
closer to where it's actually used.