Reuse snapshots of class directories on classpath
Up until now the incremental compiler would only reuse JAR
snapshots. The incremental analysis of class directories on
the classpath was inlined into the incremental analysis of the
task's output. This mean that the analysis was duplicated for
every compile task that used another task's classes folder as
an input.

This has been fixed and class directories are now treated just
like jars. This significantly reduces long term heap usage of the
incremental compiler.