Remove `TaskProvider` from task dependencies when the `Task` instance is provided
With the introduction of lazy task configuration, lots of plugins were
converted to add dependencies using `TaskProvider` instead of `Task`.
It wasn't a breaking change, but users are sometimes removing task
dependencies using the the `Task` instance. Because some dependencies
are now `TaskProvider`, it breaks the code of some users.