docs: Fix --args example for the application plugin
At least on the Windows 10 command prompt, both

    gradle run --args 'foo --bar'

and

    gradle run --args='foo --bar'

result in an "unbalanced quotes" error. Just using double quotes like

    gradle run --args "foo --bar"

results in the error

    No argument was provided for command-line option '--args'.

The only variant that works is by using double quotes and an equal sign
like

    gradle run --args="foo --bar"

Verified to also work on Git Bash for Windows and Ubuntu Linux.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>